requirements.txt
# ML0bag Vision System - Dependencies
# Core ML framework
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0

# Computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data sources
huggingface-hub>=0.20.0
datasets>=2.16.0
roboflow>=1.1.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and CLI
loguru>=0.7.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
